+++ src/main/java/com/odonto/dentisys/repository/DetalleProformaRepository.java
@@ -9,6 +9,6 @@ import com.odonto.dentisys.model.DetalleProforma;
 import com.odonto.dentisys.model.Proforma;
 
 @Repository
-public interface DetalleProformaRepository extends JpaRepository<DetalleProforma, Integer> {
+public interface DetalleProformaRepository extends JpaRepository<DetalleProforma, Long> {
     List<DetalleProforma> findByProforma(Proforma proforma);
 }
