+++ src/main/java/com/odonto/dentisys/repository/ProformaRepository.java
@@ -11,7 +11,7 @@ import com.odonto.dentisys.model.Paciente;
 import com.odonto.dentisys.model.Proforma;
 
 @Repository
-public interface ProformaRepository extends JpaRepository<Proforma, Integer> {
+public interface ProformaRepository extends JpaRepository<Proforma, Long> {
     List<Proforma> findByPaciente(Paciente paciente);
 
     List<Proforma> findByMedico(Medico medico);
